+++ lib/typescript/src/GleamContext.d.ts
@@ -8,6 +8,7 @@ export interface GleamContextValue {
     intensity: NativeProps['intensity'];
     baseColor: NativeProps['baseColor'];
     highlightColor: NativeProps['highlightColor'];
+    reduceMotion: boolean;
     registerLine: () => () => void;
 }
 export declare const GleamContext: import("react").Context<GleamContextValue | null>;